+++ packages/widgetbook/lib/src/integrations/integrations.dart
@@ -1,1 +1,0 @@
-export 'widgetbook_integration.dart';