+++ packages/widgetbook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **FEAT**: Introduce a new way to run Widgetbook via `runWidgetbook` that takes a `Config` object. ([#1597](https://github.com/widgetbook/widgetbook/pull/1597))
 - **REFACTOR**: Remove deprecated `WidgetbookCloudIntegration`. ([#1598](https://github.com/widgetbook/widgetbook/pull/1598))
+- **REFACTOR**: Rename `WidgetbookIntegration` to `Integration`. ([#1599](https://github.com/widgetbook/widgetbook/pull/1599))
 - **FIX**: Preserve component's unminified name in release builds. ([#1593](https://github.com/widgetbook/widgetbook/pull/1593))
 
 ## 4.0.0-alpha.2